1.Add countdown to Homepage
UseCountdown.jsx
Takes a countdown target date in milliseconds
and converts it's value in days, hours, minutes, seconds.

CountdownTimer.jsx
Takes a countdown target date in milliseconds.
Uses UseCountdown to convert it's value and then checks if the countdown
has reached 0.
If it has reached 0,shows a message.
If it has not reached 0 then shows the time left.

DateTimeDisplay.jsx
Takes a value and it's type (ex: days, hours, etc) and displays it.